+++ tests/Feature/Framework/MissingFile/MissingProviderFile/DomainService.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Feature\Framework\MissingFile\MissingProviderFile;
+
+final class DomainService
+{
+}